requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
gunicorn==21.2.0
lxml==4.9.3
whitenoise==6.6.0
openai==1.12.0
Pillow==10.0.0
pytube==15.0.0
opencv-python-headless==4.8.0.74
selenium==4.18.1
webdriver-manager==4.0.1
bs4==0.0.1
python-dotenv==1.0.0 